enclosure: extend the screw terminal window in the lid all the way down
This way i can solder the screw terminal to the husb238, slide the lid
down over it, and screw the screw terminal base to the lid.  Much easier
than dealing with spade terminals and stiff power wires that need to
extend out of the old window in the lid and then compress into the case.